[Waste] Create shared redirect function.

+++ perllib/FixMyStreet/App/Controller/Waste.pm
@@ -904,15 +904,10 @@ sub enquiry : Chained('property') : Args(0) {
 
     my $category = $c->get_param('category');
     my $service = $c->get_param('service_id');
-    if (!$category || !$service || !$c->stash->{services}{$service}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') unless $category && $service && $c->stash->{services}{$service};
+
     my ($contact) = grep { $_->category eq $category } @{$c->stash->{contacts}};
-    if (!$contact) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') unless $contact;
 
     my $field_list = [];
     foreach (@{$contact->get_metadata_for_input}) {
@@ -985,11 +980,7 @@ sub check_if_staff_can_pay : Private {
 
 sub bulky_setup : Chained('property') : PathPart('') : CaptureArgs(0) {
     my ($self, $c) = @_;
-
-    unless ($c->stash->{waste_features}->{bulky_enabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') unless $c->stash->{waste_features}->{bulky_enabled};
 }
 
 sub bulky : Chained('bulky_setup') : Args(0) {
@@ -1004,10 +995,7 @@ sub bulky : Chained('bulky_setup') : Args(0) {
 sub garden_setup : Chained('property') : PathPart('') : CaptureArgs(0) {
     my ($self, $c) = @_;
 
-    if ($c->stash->{waste_features}->{garden_disabled}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') if $c->stash->{waste_features}->{garden_disabled};
 
     $c->stash->{per_bin_cost} = $c->cobrand->feature('payment_gateway')->{ggw_cost};
     $c->stash->{per_new_bin_cost} = $c->cobrand->feature('payment_gateway')->{ggw_new_bin_cost};
@@ -1032,10 +1020,7 @@ sub garden_check : Chained('garden_setup') : Args(0) {
 sub garden : Chained('garden_setup') : Args(0) {
     my ($self, $c) = @_;
 
-    if ($c->cobrand->garden_current_subscription) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') if $c->cobrand->garden_current_subscription;
 
     $c->stash->{first_page} = 'intro';
     my $service = $c->cobrand->garden_service_id;
@@ -1053,14 +1038,9 @@ sub garden_modify : Chained('garden_setup') : Args(0) {
     my ($self, $c) = @_;
 
     my $service = $c->cobrand->garden_current_subscription;
-    if (!$service || $service->{garden_due}) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
+    $c->detach('property_redirect') unless $service && !$service->{garden_due};
 
-    unless ( $c->user_exists ) {
-        $c->detach( '/auth/redirect' );
-    }
+    $c->detach( '/auth/redirect' ) unless $c->user_exists;
 
     if ($c->stash->{garden_sacks} && $service->{garden_container} == 28) { # XXX Kingston sack
         my $payment_method = 'credit_card';
@@ -1106,14 +1086,8 @@ sub garden_modify : Chained('garden_setup') : Args(0) {
 sub garden_cancel : Chained('garden_setup') : Args(0) {
     my ($self, $c) = @_;
 
-    if (!$c->cobrand->garden_current_subscription) {
-        $c->res->redirect('/waste/' . $c->stash->{property}{id});
-        $c->detach;
-    }
-
-    unless ( $c->user_exists ) {
-        $c->detach( '/auth/redirect' );
-    }
+    $c->detach('property_redirect') unless $c->cobrand->garden_current_subscription;
+    $c->detach( '/auth/redirect' ) unless $c->user_exists;
 
     $c->forward('get_original_sub', ['user']);
 
@@ -1645,6 +1619,18 @@ sub uprn_redirect : Path('/property') : Args(1) {
     my $result = $echo->GetPointAddress($uprn, 'Uprn');
     $c->detach( '/page_error_404_not_found', [] ) unless $result;
     $c->res->redirect('/waste/' . $result->{Id});
+}
+
+sub property_redirect : Private {
+    my ($self, $c) = @_;
+    $c->res->redirect('/waste/' . $c->stash->{property}{id});
+}
+
+sub label_for_field {
+    my ($form, $field, $key) = @_;
+    foreach ($form->field($field)->options) {
+        return $_->{label} if $_->{value} eq $key;
+    }
 }
 
 __PACKAGE__->meta->make_immutable;
